+++ mcww/ui/misc/optionsUI.py
@@ -177,7 +177,7 @@ class OptionsUI:
                         self._components.forceShowBatchCount = gr.Checkbox(label='Show "Batch count" parameter even if workflow doesn\'t have any seeds')
                         self._components.hideHomepagesInFooter = gr.Checkbox(label='Hide homepage links in the footer (MCWW, Gradio, ComfyUI)')
                         self._components.showToggleDarkLightButton = gr.Checkbox(label='Show "☀️/🌙" button for changing dark/light theme (still functional when hidden)')
-                        self._components.useCustomContextMenu = gr.Checkbox(label="Use custom context menu instead of browser's context menu. (Everything except text boxes)")
+                        self._components.useCustomContextMenu = gr.Checkbox(label="Use custom context menu. Works on RMB, long finger touch, 2 fingers tap")
                 with gr.Tab("Theme"):
                     with gr.Group():
                         self._make_primaryColorOptions()